Random walk through strings → URLs
Hypothesis: handling URLs as a proper URL struct would be safer than
handling them as strings. This commit is an exploration to see how
invasive it would be to refactor.

tl;dr this won't go into production, it's hundreds of changes, with
hundreds to go.

I started by choosing some function, changing a string parameter to take
a URL, fix the red squiggles, and repeat.

I think the end state of using URLs is desirable, but getting there is
very tedious and perhaps error-prone. On the upside, I believe it would
be safer, more testable, and would on balance reduce lines of code and
error checks.

It might result in more allocations. Not sure! Could go either way.

Another concern is mutability. The *url.URL struct is typically
passed as a reference, and so a callee could modify the caller's value.
Some sort of immutable URL might be helpful.